Fixes to yAxis rendering and cleanup from prior commit
This improves the Y axis rendering, fixing issues introduced in the previous commit as well as providing better range calculation logic.
Tests are still failing as part of this commit, although these changes are believed to be good more validation is needed.  The next commit should cleanup any remaining issues and fix the unit tests.