Add `LabelSkipCount` to the YAxisOption, plus more tweaks
This adds the `LabelSkipCount` configuration as a way to reduce the amount of y axis labels, while maintaining the same `LabelCount` number of horizontal lines.

As part of testing around this change it was found that the Unit parameter was not handled in the best way possible.  A long comment in charts.go decribes the circular relationship between the axis range and label count (particularly if a unit is configured).