Add new "vivid" light and dark themes and remove Set functions from ColorPalette
These themes are a nice addition to the default, providing another light and dark color choice with red, yellow and other bright colors up front on the series.  These themes are useful for any data where the initial items in the series should have the most emphasis.

In addition the ColorPalette interface was changed to remove Set functions so that these structs will always only be read only.