Update capped bar graph implementation to use arc draw
This brings some of the logic from commit 9b7634c2c230a0a1907c6e99be6bdea11f811a9e in the upstream project to allow a "RoundedRect" draw option.

The upstream implementation was updated to conditionally set if the top or bottom should be rounded, for our configuration we only allow rounding the top of the bars (other config changes were not brought in).

Benchmarking shows this to be about the same, and the SVG size is slightly larger with this arc draw method.  However I believe this is superior to needing to re-draw the background if the circle extends past the axis.  That method was awkward and given certain text rotations could result in cliping (I had planned to re-order the rendering to fix that, but I prefer this option).