Change Series Styling and Themes
This change removes the ability to configure the styling at the series level.

As seen in this change and discussed in issue #14, these style options were not very broadly recognized.

Instead the goal is to make theme customization easier. As a potential replacement to this configuration the Themes can now accept series and background color changes.

If we do decide that we want to bring back per-series styling, I think we should make those top level fields so that we only introduce fiels which are broadly recognized.