Remove `Style` from alias, and create `chartdraw.FontStyle`
This change removes the `FontSze`, `FontColor`, and `Font` from the `chartdraw.Style` struct and instead moves them to a `chartdraw.FontStyle`.

This is the first change since forking `wcharczuk/go-chart` that will make a transition require additional changes.  However I believe this is an important change.

Using the broad `Style` struct makes it non-obvious what values are actually being used.  In addition these font configurations are a common need, making an obvious choice for needing to bundle together.

By removing the `Style` from `charts`, we instead are elevating these fields to be part of the public configuration.  `Style` is now only used for rendering, and not part of the public API (unless using the `wcharczuk/go-chart` fork directly).