Legend and Title Offset Config Changes
This change moves the `Top` and `Left` configuration options for a Legend or Title under the `OffsetStr` struct.
In cases where only `Left` was being set with `PositionLeft` / `PositionCenter` / `PositionRight`, now `OffsetLeft` / `OffsetCenter` / `OffsetRight` can be used to set the `Offset` field.
In addition sub-structs in configurations should consistently have `WithX` functions to allow easy modifications or a builder pattern.