Improve default value formatting
This commit improves and exposes the formatting logic used by default for value formatting.

The behavior remains that by default the values over 1,000 will be reduced and suffixed with a `k`, `M`, `G`, `T` as appropriate.  It also remains that by default up to 2 decimals will be specified.

However this value previously was truncated rather than rounded.  Now if a decimal or other truncation of the value occurs rounding will keep values as close as possible.

In addition this functionality is exposed in our API, the default formatting available through the function `FormatValueHumanizeShort`.  Also exposed is `FormatValueHumanize` which will provide the rounding and comma logic without the truncation and unit suffix.  Both these functions allow you to specify how many decimals should be included, and if they should be forced (suffixing zeros to meet the decimal length if necessary).