chartdraw: Fix possible race condition loading default font
`chartdraw` tests were updated to be in parallel.  Doing so exposed a possible race condition in `font.go`, although it only would result in minimal extra load of the font.
To fix this the implementation from the root `font.go` was moved to `chartdraw`, and now the root `font.go` just references the `chartdraw` implementation.  This has slightly nicer ergonomics in that the error handling is no longer necessary.