De-duplicate, fix, and improve color parsing
There was a very similar color parsing implementation in `color.go` and `chartdraw/drawing/color.go`.  After benchmarking it was found that `color.go` was faster, but also incorrectly handled the alpha channel.

This commit updates the implementation in `chartdraw/drawing/color.go` to use this simpler and more flexible regex pattern.  Allowing both rgb and rgba formats to be parsed by `ColorFromRGBA` in a more performant way.  With this change the implementation in `color.go` was removed to instead defer to the improved `chartdraw/drawing/color.go` implementation.  PraseColor has been elevated to the `charts` package so this functionality can be easily found.