Fix: Only add extra tick mark when centering labels
This partially addresses issue #8, reducing the conditions that may have an extra label to charts using BoundaryGap (enabled by default) and with a large dataset.

The logic change in `isTick` should be identical to the prior for-loop implementation, however faster and simpler.  The fix is otherwise addressed by removing the unconditional `+1` to the label count and instead only incrementing the count if we are centering the labels (and thus need an additional tick mark to center against).

Large datasets face an issue in how they compare the labels and their positions, since the label may center against a non-rendered tick position.