line_chart: Default SymbolShow will only be enabled for smaller datasets
Very dense datasets may have the symbol collide or make the graph look messy.
If the user does not specify the behavior, we will auto-disable it when data points are 100 or greater.

In the future we may make this based off the render space, but this is likely a simple and decent improvement.