Improve Title and Legend positioning and chart spacing
This change was initially motivated to add `bottom` as a value that can be specified for the legend or title when rendering on the bottom is desired.

When adding that feature it was discovered that the title and legend spacing was inconsistent.  For example if no title was specified the chart would extend into the legend.  This change attempts to make that spacing more predictable and controllable:
  * If a title and legend exist on the top, they will attempt to share the space as done previously
  * If only a title exists on the top they will be given the space needed
  * If only a horizontal legend exists on the top, by default it is given space, but the prior behavior can be achieved by setting `OverlayChart` to `*true` in the `LegendOptions`
  * If a title or legend is set at the bottom, the chart space at the bottom will be given in a similar way as done on the top

In general the change here is pretty minor, but I believe this will be more predictable for the user, and given our tests overall this produced better spaced charts.

In the future we may enable `OverlayChart` to accept a `*false` to provide different behavior for vertical legends.  This is why I felt a pointer may be better than a simple bool field.