line_chart: Add `StrokeSmoothingTension` config option
This added option allows you to specify a value between 0 and 1 to curve and smooth the line.

As commented this will result in the line being slightly drifted from the set points in order to accomplish this smoothing.  But for datasets with minimal variation, or wide spacing, between points it can provide a nice visual effect.