Change how SeriesList are built
This removes from the public API NewSeriesListDataFromValues and NewSeriesFromValues.  Instead of optionally passing in a chart type it's now expected to call a function specific to the chart type: NewSeriesListLine, NewSeriesListBar, NewSeriesListHorizontalBar, NewSeriesListPie, NewSeriesListRadar, or NewSeriesListFunnel.

NewPieSeriesList and NewFunnelSeriesList previously existed, so this API makes the construction more uniform for all chart types (note that these have been renamed to have a `NewSeriesList` prefix to the function name).

While examining the Options struct passed in (which has now been adopted for all series types), it was noted that Pie rendering almost always passed in a SeriesLabel to enable rendering the labels.

The examples demonstrate these API changes well, hopefully this API is more concise and easier to learn.