Remove `error` return on `NewPainter`
This removes the error return on `chartdraw.RendererProvider`.  Currently the only error return was for a PNG case that was due to a type checking that would have been from a bug.  Instead the function signatures were changed so the type check was no longer necessary, making the error result no longer needed.

The above change was able to cascade until `charts.NewPainter` no longer needed to return an error result either.