Painter API change away from `New` functions and `Render()`
This commit removes the need for external users to be aware of Renders (`Renderer` type made private in a prior commit).  Instead if a user wants to render charts on a Painter they can invoke the relevant function on the painter:
* NewBarChart -> Painter.BarChart
* NewHorizontalBarChart -> Painter.HorizontalBarChart
* NewFunnelChart -> Painter.FunnelChart
* NewLineChart -> Painter.LineChart
* NewPieChart -> Painter.PieChart
* NewRadarChart -> Painter.RadarChart
* NewTableChart -> Painter.TableChart

In additional `TableOptionRender` was renamed to `TableOptionRenderDirect` and `TableRender` was renamed to `TableRenderValues`, both to provide clarity because these render functions are unique compared to others.