Style and minor performance improvements
This change updates `golangci-lint` to include a range of linters not previously used. The goal of this is to help maintain a consistent style across the project.

Style changes were made as found from the linter, in addition to some manual review.

Performance gains were from preallocating slice sizes and also reducing the usage of `fmt.Sprintf`.