Add `BorderWidth` to the `LegendOption`
This option allows a border to be drawn around the legend by setting a non-zero value.

Implementing this feature highlighted an error in how the returned Box was being defined.  This fixes that error, but as seen in the lack of test changes the overal chart rendering remains the same.

As noted by a couple TODO notes, we may want to improve this option in the future:
* Right now the border is always black, we could add this to the Theme or the LegendOption. I could not decide where to add this yet so right now it's only black.
* It would be nice to draw the background behind the legend. That right now would require us to calculate the dimensions up front, so this is for now being left as a future improvement.