PNG rendering performance improvements
This commit expands the `Painter` `benchmark` to be per-chart type instead of matching the prior multi-chart behavior.

In addition, a significant performance gain is made for when filled rectangles are rendered to a PNG.  This provides a speed benefit to every chart when the background is filled, but it provides a more significant gain for bar charts or when more boxes must be rendered.