Add `BarMargin` option to `Bar` and `HorizontalBar` charts
This option allows users to control the space between each bar within a series (and in effect the space between series).

In addition a bug was discovered where the Painter API may not correctly scale the chart when Markpoints are used. This bug is another permutation of #30 that was undiscovered, and fixed by moving this logic under `defaultRender`.