drawing: Implement historic TODO notes for handling `Close()` in flattener
There were TODO notes to store the `MoveTo` location and correctly handle the `Close()`.  Prior to this change this was being handled in the `flattener` with an extra `LineTo` call.  With this change we ensure `Close()` is implemented in all interface implementors and now the extra `LineTo` is no longer necessary.