Add `LineChartOption.StackSeries` option
This option layers the series on top of each other. This can be a great visualization when the sum of the values is an important factor to show in the line. Similar to Bar charts this option has documented options which are not supported or behave different.

This option was also added to `ChartOption` so that both API's can access this behavior.