chartdraw: Mark default fields as deprecated
These fields are not expected to be used externally. I want to ensure useless fields and functions are internal so the public API is easier to adopt for new users.

If any of these fields are useful, please open an issue with how it's used so we can maintain the field publicly. The goal here is still to maintian the `chartdraw` api, and only make it easier to navigate and use.