Add `NewFontStyleWithSize(int)`
Creating a FontStyle just to set a size is a common operation.  This function reduces the boilerplate for doing that action.  It can also be combined with `With` (for example `WithColor`) to easily build other variations as well.

This commit adopts this API, as well as the `NewBox` and `NewBoxEqual` added previously, significantly reducing the project line count.