pie_chart: Fix per-series radius configuration
Prior to this change the last radius in the series was all that was used.
Not it's possible to set a specific series / pie slice to a different radius from the rest. However this does require setting the radius of all slices in the series.

This is considered a bug fix based off what the configuration currently allows. In the future we want to allow a chart radius and a series radius independently.