util: Remove internal sumInt | chartdraw: SumInt handle overflows
As documented in the comment, this was also implemented in chartdraw. Now that the code is colocated we can easily improve the chartdraw implementation to handle the overflow / underflow conditions.

By doing this it was also discovered that charts.sumInt had a bug in the underflow check. This was discovered through the unified unit tests.