Series Type to match Chart Type
This breaks the single `Series` into a series per-chart type.

This helps make the API more obvious, by knowing that available fields are all potentially useful if desired. It also allows the `Value` field to store a single value for charts that only support a single series value (Funnel and Pie). In addition future changes easier as we expand support for series specific marks and adjustements.

The result is a significant increase in the code and field duplication within series.go. However this is relatively simple code, making it easily worth the improved external API this offers.