feat: Add Scatter chart type and SeriesTrendLine
This chart type is very similar to Line charts. The notable design difference is that for a given X position it's possible to have multiple values. This is why the ScatterSeries.Values is `[][]float64`, the second dimension does not have to be the same size, but offers a way to specify multiple points at an X position.

It was considered to put all needed options in the line chart implementation (or even allowing the `LineStrokeWidth` to be set to `0`). However I don't want to complicate line charts to handle scatter charts, and similar I want scatter charts to allow easy future expansions. The possibility for multiple points at an X position is a notable reason for why adapting this in the line chart would be difficult.

Also added (and able to be used in Scatter charts) is the ability to add a Trend line. This draws a line based off the series data to show Linear, Cubic, and Moving Average trends on scatter charts.

Configuration is similar (but slightly different) from SeriesMarkLines, constructed through `NewTrendLine`.