Remove YAxisOption.AxisColor and instead configure in the Theme
Themes should provide a central place for coloring your charts. It was inconsistent and confusing that the theme only defined a single axis color, with `YAxisOption` being a unique case where the Color was directly set.

As seen in the unit tests this was an easy method for setting the a Y Axis to match the series color (Particularly when two Y-Axis's are used, future improvement coming). For this reason the Theme offers a `WithYAxisSeriesColor` function so you can set the YAxis to match a series (typically 0 or 1 depending on the axis index).