feat: Add the ability to configure a border in the TitleOption
This change is similar to already supported in the LegendOption, adding the float64 field `BorderWidth` to define a border around the title. In addition the Theme was expanded to allow a configuration for the color of this border (similar to Legend).

Echarts was expanded to use this configuration field as well.