fix: Fix for charts with right y-axis only
This change fixes the position of the x-axis and data values when the y-axis is on the right side of the chart only.
This also improves the handling of the Position field and axis index (for example if desired you could have the right y-axis at position 0 to get the split lines drawn on the right axis).