fix: Better handling for variable x-axis sizes
Axis spacing is challenging because all the axis depends on other axis. For example Y-Axis need to be shifted up above the X.  And the X is squished between any rendered Y-Axis (making it's width shorter).

Prior to this change the x-axis size was defined as a constant. In this change we:
* Redo axis.go's rendering logic so that the painter padding and sizing logic is more intuitive.
* Instead of measuring the x-axis title and adding the constant size, we render the x-axis on a matching painter and measure it's height.

Overall this makes the code a bit more flexible (different stroke widths, taller tick marks, more padding options to the title, etc).