fix: Better handling for variable x-axis sizes
Axis spacing is challenging because all the axis depends on other axis. For example Y-Axis need to be shifted up above the X.  And the X is squished between any rendered Y-Axis (making it's width shorter).

Prior to this change the x-axis size was defined as a constant. The constant does simplify the problem, but is less flexible for variations in the stroke sizes, tick marks, label sizes, and recently titles.

At a high level this change is focused on:
  1. Redo axis.go's rendering logic so that the painter padding and sizing logic is more intuitive.
  2. Instead of measuring the x-axis title and adding the constant size, we render the x-axis on a matching painter and measure it's height.
  3. Vector painting sizing switched from int cast to now be rounded.

Details on each:
  1. A bug after introducing axis titles resulted in focusing in axis.go's implementation. These changes have also allowed us to be a bit tighter on how much space the x/y-axis needs.
  2. A flexible x-axis height is very challenging, even if the axis can be rendered several times to the same dimensions, it's still critical that the painter coordinates relative to each other exactly match. The careful math and `painterPrePositioned` option were critical to making this work.
  3. The vector renderer measure and positioning were colliding in some cases with the tighter axis rendering. Switching to round and ceil as needed were a clean fix to keep the logic in axis.go much simpler with as few constant tweaks as possible.

Because axis's are rendered in almost every chart, this has resulted in a lot of unit test updates. SVG changes were reviewed by hand, in general the axis uses a little less space, and if disabled a lot of space is reclaimed.