refactor: Centralize label logic within `axisRange` + Deprecate YAxisOption.Formatter
`YAxisOption.Formatter` is deprecated in favor of `ValueFormatter` which can enable similar functionality. This extra method was found to be more combersome in the refactor.

The primary goal of the refactor is to bring all the axis label measuring, and count logic from `defaultRender` + `axisPainter` + `axisRange` into entirely `axisRange`. Choosing a good number of labels is difficult:
* It depends on the physical space available to render
* It depends on the data range itself
* The user may specify things like custom units, or specific label counts
* Large or dense labels necitate a need to render less labels for visual space
* We want to choose increments that result in human friendly numbers

Right now we do many of these very good, but there are combinations which should be improved. In order to do that we need to centralize this logic so the interactions from each point can be balanced out. `axisRange` was choosen for the central place for this logic because it was already doing much of the range padding logic, and could be done as a calculation step before the axis is rendered.

Additional fixes found and included:
* fixed reducing y-axis label counts
* Horizontal bar lengths are now more accurate to the x-axis labels (very minor)