fix: Fix for axis scales to be more human friendly
This commit embraces the prior refactor to produce more human friendly axis labels and intervals. Although there were several changes to accomplish this, the most significant was measuring some optimistic labels early so that we can let padRange / friendlyRound, and new logic around unit handling do their logic without late interference once the labels are actually measured.

This results in potentially needing to generate and measure the labels twice, but is worth the tradeoff for the resulting axis values.

Included are the unit tests from the prior commits changes, which included a careful review (and a bug fix for rotated axis labels as part of that review).