fix: Adjust too dense labels when boundary gap is disabled
The changes in v0.5.2 resulted in category axis being potentially more dense. Normally this results in a cleaner look, but when the boundary gap is disabled the last two labels can collide.

This change increases the spacing for the category axis to be closer to the behavior in v0.5.1, though still often better spaced.