fix: Further tick reduction for dense datasets
This contines the change done in v0.5.3, further reducing the tick count. With the last change the difference in tick to label count can make it hard to read.

This change enforces a ratio of tick to label counts with the goal of making the label to tick position clearer to read.

Unfortunately the fallout from v0.5.2 has required several iterations to get the readability right.