fix: Additional fixes with extra tick marks
After further review this problem is deeper than realized. I have struggled to arrive at an algorithm that will reduce the tick marks in a way that will reliably line up with the labels in all cases.

For that reason this change will just reduce the tick count to match the label count if the tick counts exceed 2 times the label count.