Deprecate `Font` on ChartOption structs
This field is generally used on the series labels, but not adopted for axis's, title, or legend. It may be confusing, and since this can be specified using the SeriesLabel directly, it's more obvious if we go ahead and deprecate this.

On the SeriesList we now support `SetSeriesLabels` so that the series Font, Show, Formatting, or other configuration can be uniformly set across the series.