fix: raster fix for making smooth large circles
Using `QuadCurveTo` looks fine for small smaler arc's, but when the circle is large it starts to take a more square shape due to the limited QuadCurveTo calls.

It was considered to make the curve shorter, but it turns out using `ArcTo` is a much simpler solution to producing consistently smooth circles.