feat: Add doughnut chart implementation and `SegmentGap`
This change brings in a new `doughnut` chart type.

The doughnut chart is being created by rendering a pie chart first, then putting a circle, and center data afterwords. Most of the doughnut chart implementation is the support for labels on the inside circle. Label distribution turned out to be a very challenging problem, and still only works when the series count is reasonably low.

As part of this the pie chart implementation was carefully reviewed, and given some small refactors around the sector and label placement, as well as break apart and move some utility functions.