fix: Small stacked series always shown in stack
Prior to this change the value range may result in the range minimum value matching the series miniumum value. This results in series bars / lines which normally would meet the X axis, but in a stacked series it results in the series just being hidden.

In this change, when a series is stacked we just subtract one from the mininum value, resulting in even the smallest series representing at least "one" value worth of space.