fix: Fix stacked category axis spacing (bar charts)
Our prior logic in `range` for producing a category axis would use the series names for labels. This is not correct, instead the axis should represent the data span.

This was most obvious in bar charts when the series is stacked. It would show the bar as being pushed to the edge for each series (that was instead stacked instead of beside).