fix: Fix for preferring smaller label counts when unit is used
This change fixes a regression introduced in the prior change (v0.5.12). The improved unit label count / span logic failed to have a preference for more labels. Resulting in cases where two labels are equally good as charts where every value is labeled.
This updates the conditional to prefer higher label counts when the delta from the padded label count is the same.