fix: more robust SVG emoji meansurements
There is always an inconsistency due to our library font being potentially different than the browser or rendering app fonts.
This is particularly notable with emoji's which are not in our font, resulting in a missing character that is typically much smaller.

This change improves that by detecting this condition and instead creates a square space for the emoji.  Overall this is much closer, but not perfect.