feat: Add `notosans` and `notosans-bold` fonts
This adds to the chart library these additional fonts in the default set. Notosans provides better international, symbol, and emoji support. The addition of a `bold` format (which is actually `ExtraBold`) also provides more styling options.

These fonts are provided in a compressed format to minimize the library storage size. They are also loaded lazily in order to minimize the in-memory impacts to only what's used.