feat: Allow individual series label customization
Prior to this change all data labels had to be enabled or disabled, as well as with a matching style. This left mark points to be the only way to highlight specific values.

With this change a `SeriesLabelFormatter` can now be set on the `SeriesLabel` struct. This function allows for selective hidding of values (by returning an empty string), as well as selective changes to the text style or a background behind the label.

This is designed for emphasizing specific values, and with that `LabelFormatterThresholdMin`, `LabelFormatterTopN`, `LabelFormatterProgressiveGreenRed`, and `LabelFormatterProgressiveColor` were added as helpers to easily highlight upper values.

`SeriesLabel.FormatTemplate` has been deprecated to be removed in `v0.6.0`.  The new `LabelFormatter` function is a more powerful alternative.