feat: Improved emoji and symbol support
The recently added Noto Sans has several symbols and emoji's not within Roboto.  While Roboto has several mathmatical symbols not within Noto Sans.
With this change, regardless of what font is used, if a symbol is missing in the font it will fallback to non-bold bundled font variants.
With SVG's this results in more accurate font measuring (since symbols will be shown even if we can't measure them).

Much of the font core had to be moved to the `chardraw/drawing` package so that the same central loading can be utilized by the png loading.