feat: Candlestick / Klines chart type with Patterns
This change adds a new chart type: "candlestick"

These charts track the open, close, high, and low values typically used in finchances.
We are including patterns which are often used to highlight specific points on this chart type.

This chart also prototypes some API changes that are being considered for v0.6.