fix: ChartOption preserve series colors when combining chart types
In v0.5.0 we updated the series structs to be chart type specific. This resulted in the filter operation completely removing different types from the series slice.

The above change resulted in the bug of after series filtering each chart type would be referencing the series starting from index 0.

This change fixes that behavior by adding an `absThemeIndex` to series types which can be combined. This internal field is set when sourced from a `GenericSeries` to ensure the original series ordering is communicated to the chart rendering.

Tests were added to cover this behavior, which also highlighted that our chart rendering order could be improved for readability as well.